requirements.txt
requests
chromadb
gitpython
python-dotenv
streamlit
pandas